Updated laptop project. FIREBASE STILL NOT WORKING. However i made all objects primitive type to conform to firebase json storeage..

+++ app/src/main/java/app/reservation/acbasoftare/com/reservation/Recycleview/RVAdapter.java
@@ -78,7 +78,7 @@ public class RVAdapter<E> extends RecyclerView.Adapter<RecyclerView.ViewHolder> 
         } else {//do service
             DecimalFormat df = new DecimalFormat("$ ###,##0.00");
             SalonService s = (SalonService) list.get(position);
-            rb.setText("Service: " + s.getName() + "\n" + "Price: " + df.format(s.getPrice().doubleValue()) + "\nService Duration: " + s.getDuration());
+            rb.setText("Service: " + s.getName() + "\n" + "Price: " + df.format(s.getPrice()) + "\nService Duration: " + s.getDuration());
         }
     }
